Bit length set extracted into a separate file
